T:O(logn + logx) S:O(1)

+++ daily_challenge/3133_minimum_array_end.cpp
@@ -7,7 +7,7 @@ public:
       if (x & 1) {
         out += 1 << i;
       } else {
-        out += (n & 1) << i;
+        out += (size_t)(n & 1) << i;
         n >>= 1;
       }
     }
